refactor(shell): move release loader to shared lib
