+++ apps/web/app/app/(shell)/releases/ReleasesRoute.tsx
@@ -1,37 +1,44 @@
-import { notFound, redirect } from 'next/navigation';
+import { redirect } from 'next/navigation';
 import { APP_ROUTES } from '@/constants/routes';
 import { PageErrorState } from '@/features/feedback/PageErrorState';
 import { getCachedAuth } from '@/lib/auth/cached';
 import { captureError } from '@/lib/error-tracking';
-import { getAppFlagValue } from '@/lib/flags/server';
 import { queryKeys } from '@/lib/queries';
 import { HydrateClient } from '@/lib/queries/HydrateClient';
 import { getDehydratedState, getQueryClient } from '@/lib/queries/server';
 import { loadReleaseMatrix } from '@/lib/releases/release-matrix-loader';
 import { getDashboardShellData } from '../dashboard/actions';
-import { LibraryPageClient } from './LibraryPageClient';
+import { ReleasesClientBoundary } from '../dashboard/releases/ReleasesClientBoundary';
+import { ReleasesPageClient } from '../dashboard/releases/ReleasesPageClient';
 
-export const runtime = 'nodejs';
-
-export default async function LibraryPage() {
+/**
+ * Releases page — client-first with shell-only server work.
+ *
+ * Auth check via getCachedAuth (Clerk JWT, no DB) runs first. The shared shell
+ * and /app route warm the release matrix cache ahead of navigation, so this
+ * page keeps warm transitions instant while still hydrating release data on
+ * direct visits, refreshes, and bookmarks.
+ */
+export async function ReleasesRoute() {
   const { userId } = await getCachedAuth();
-  const libraryEnabled = await getAppFlagValue('SHELL_CHAT_V1', {
-    userId,
-  });
-
-  if (!userId || !libraryEnabled) {
-    notFound();
+  if (!userId) {
+    const signInParams = new URLSearchParams({
+      redirect_url: APP_ROUTES.RELEASES,
+    });
+    redirect(`${APP_ROUTES.SIGNIN}?${signInParams.toString()}`);
   }
 
+  // Shell data is cached from the shell layout (same request) — resolves instantly.
   const dashboardData = await getDashboardShellData(userId);
+
   if (dashboardData.dashboardLoadError) {
-    await captureError(
-      'Dashboard data load failed on library page',
+    void captureError(
+      'Dashboard data load failed on releases page',
       dashboardData.dashboardLoadError,
-      { route: APP_ROUTES.LIBRARY }
+      { route: APP_ROUTES.RELEASES }
     );
     return (
-      <PageErrorState message='Failed to load library data. Please refresh the page.' />
+      <PageErrorState message='Failed to load releases data. Please refresh the page.' />
     );
   }
 
@@ -49,10 +56,10 @@ export default async function LibraryPage() {
       });
     } catch (error) {
       void captureError(
-        'Release matrix prefetch failed on library page',
+        'Release matrix prefetch failed on releases page',
         error,
         {
-          route: APP_ROUTES.LIBRARY,
+          route: APP_ROUTES.RELEASES,
         }
       );
     }
@@ -60,7 +67,9 @@ export default async function LibraryPage() {
 
   return (
     <HydrateClient state={getDehydratedState()}>
-      <LibraryPageClient />
+      <ReleasesClientBoundary>
+        <ReleasesPageClient />
+      </ReleasesClientBoundary>
     </HydrateClient>
   );
 }
