+++ apps/web/tests/lib/queries/gc-time-coverage.test.ts
@@ -13,7 +13,7 @@ vi.mock('@tanstack/react-query', async () => {
   };
 });
 
-vi.mock('@/app/app/(shell)/dashboard/releases/release-matrix-loader', () => ({
+vi.mock('@/lib/releases/release-matrix-loader', () => ({
   loadReleaseMatrix: vi.fn(),
 }));
