+++ apps/web/tests/unit/app/dashboard-metadata.test.ts
@@ -43,7 +43,7 @@ vi.mock('@/app/app/(shell)/dashboard/releases/actions', () => ({
   }),
 }));
 
-vi.mock('@/app/app/(shell)/dashboard/releases/release-matrix-loader', () => ({
+vi.mock('@/lib/releases/release-matrix-loader', () => ({
   loadReleaseMatrix: vi.fn().mockResolvedValue([]),
 }));
