+++ apps/web/tests/unit/calendar/calendar-page.test.tsx
@@ -62,7 +62,7 @@ vi.mock('@/app/app/(shell)/dashboard/actions', () => ({
   getDashboardShellData: mocks.getDashboardShellData,
 }));
 
-vi.mock('@/app/app/(shell)/dashboard/releases/release-matrix-loader', () => ({
+vi.mock('@/lib/releases/release-matrix-loader', () => ({
   loadReleaseMatrix: mocks.loadReleaseMatrix,
 }));
